% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_offspring_members.R
\name{get_offspring_members}
\alias{get_offspring_members}
\title{get_offsprint_member}
\usage{
get_offspring_members(group_label, deepth = 0)
}
\arguments{
\item{group_label}{(character) A string, it can be name of genus, family, order, subclass, class, etc.}

\item{deepth}{(integer) A integer}
}
\value{
(character) A table of taxonomic scheme of the corresponding groups
}
\description{
get_offsprint_member
}
