% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parent_lineage.R
\name{get_parent_lineage}
\alias{get_parent_lineage}
\title{get_parent_lineage}
\usage{
get_parent_lineage(group_label, deepth = 0)
}
\arguments{
\item{group_label}{}

\item{deepth}{}
}
\description{
get_parent_lineage
}
